+++ tests/testthat/test-estimate_contrasts_methods.R
@@ -173,12 +173,7 @@ test_that("estimate_contrasts - random effects, single by/contrast", {
     estimate_contrasts(model, contrast = "gear"),
     regex = "Could not calculate"
   )
-  expect_message(
-    {
-      estim <- estimate_relation(model, by = "gear")
-    },
-    regex = "Standard errors are probably not reliable"
-  )
+  estim <- expect_silent(estimate_relation(model, by = "gear"))
   out <- estimate_contrasts(estim, contrast = "gear")
   expect_identical(dim(out), c(3L, 8L))
   expect_named(